command: legacy, move all deprecated commands to legacy
