+++ pkg/cmd/sctool/sctool.go
@@ -9,25 +9,25 @@ import (
 
 	"github.com/scylladb/scylla-manager/pkg/command/backup"
 	"github.com/scylladb/scylla-manager/pkg/command/backup/backupdelete"
-	"github.com/scylladb/scylla-manager/pkg/command/backup/backupfiles"
 	"github.com/scylladb/scylla-manager/pkg/command/backup/backuplist"
 	"github.com/scylladb/scylla-manager/pkg/command/backup/backupvalidate"
 	"github.com/scylladb/scylla-manager/pkg/command/cluster/clusteradd"
 	"github.com/scylladb/scylla-manager/pkg/command/cluster/clusterdelete"
 	"github.com/scylladb/scylla-manager/pkg/command/cluster/clusterlist"
 	"github.com/scylladb/scylla-manager/pkg/command/cluster/clusterupdate"
+	"github.com/scylladb/scylla-manager/pkg/command/legacy/backupfiles"
+	"github.com/scylladb/scylla-manager/pkg/command/legacy/task/taskdelete"
+	"github.com/scylladb/scylla-manager/pkg/command/legacy/task/taskhistory"
+	"github.com/scylladb/scylla-manager/pkg/command/legacy/task/tasklist"
+	"github.com/scylladb/scylla-manager/pkg/command/legacy/task/taskprogress"
+	"github.com/scylladb/scylla-manager/pkg/command/legacy/task/taskstart"
+	"github.com/scylladb/scylla-manager/pkg/command/legacy/task/taskstop"
+	"github.com/scylladb/scylla-manager/pkg/command/legacy/task/taskupdate"
 	"github.com/scylladb/scylla-manager/pkg/command/repair"
 	"github.com/scylladb/scylla-manager/pkg/command/repair/repaircontrol"
 	"github.com/scylladb/scylla-manager/pkg/command/resume"
 	"github.com/scylladb/scylla-manager/pkg/command/status"
 	"github.com/scylladb/scylla-manager/pkg/command/suspend"
-	"github.com/scylladb/scylla-manager/pkg/command/task/taskdelete"
-	"github.com/scylladb/scylla-manager/pkg/command/task/taskhistory"
-	"github.com/scylladb/scylla-manager/pkg/command/task/tasklist"
-	"github.com/scylladb/scylla-manager/pkg/command/task/taskprogress"
-	"github.com/scylladb/scylla-manager/pkg/command/task/taskstart"
-	"github.com/scylladb/scylla-manager/pkg/command/task/taskstop"
-	"github.com/scylladb/scylla-manager/pkg/command/task/taskupdate"
 	"github.com/scylladb/scylla-manager/pkg/command/version"
 	"github.com/scylladb/scylla-manager/pkg/managerclient"
 	"github.com/spf13/cobra"
