+++ pkg/command/legacy/task/taskdelete/cmd.go
