+++ pkg/command/legacy/task/taskprogress/res.yaml
